+++ packages/facility-server/app/routes/apiv1/patient/patientVaccine.js
@@ -1,6 +1,5 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
-import { subDays, addDays } from 'date-fns';
 import Sequelize, { Op, QueryTypes } from 'sequelize';
 import config from 'config';
 
@@ -12,7 +11,7 @@ import {
   VISIBILITY_STATUSES,
 } from '@tamanu/constants';
 import { NotFoundError } from '@tamanu/shared/errors';
-import { getCurrentDateString, toDateTimeString } from '@tamanu/shared/utils/dateTime';
+import { getCurrentDateString } from '@tamanu/shared/utils/dateTime';
 
 export const patientVaccineRoutes = express.Router();
 
@@ -26,106 +25,6 @@ const asRealNumber = value => {
   }
   return num;
 };
-
-const mockDueDate = index => {
-  let dueDate;
-
-  switch (true) {
-    case index === 0:
-      dueDate = toDateTimeString(addDays(new Date(), 30));
-      break;
-    case index === 1:
-      dueDate = toDateTimeString(addDays(new Date(), 10));
-      break;
-    case index === 2:
-      dueDate = toDateTimeString(addDays(new Date(), 1));
-      break;
-    case index === 3:
-      dueDate = toDateTimeString(subDays(new Date(), 10));
-      break;
-    default:
-      dueDate = toDateTimeString(subDays(new Date(), 60));
-      break;
-  }
-
-  return toDateTimeString(dueDate);
-};
-
-// Todo: Update in NASS-1146
-patientVaccineRoutes.get(
-  '/:id/vaccineSchedule',
-  asyncHandler(async (req, res) => {
-    req.checkPermission('list', 'PatientVaccine');
-
-    const results = await req.db.query(
-      `
-      SELECT
-        sv.id
-        , max(sv.category) AS category
-        , max(sv.label) AS label
-        , max(sv.schedule) AS schedule
-        , max(sv.weeks_from_birth_due) AS weeks_from_birth_due
-        , max(sv.vaccine_id) AS vaccine_id
-        , max(sv.visibility_status) AS visibility_status
-        , count(av.id) AS administered
-        FROM scheduled_vaccines sv
-        LEFT JOIN (
-          SELECT
-            av.*
-          FROM
-            administered_vaccines av
-            JOIN encounters e ON av.encounter_id = e.id
-          WHERE
-            e.patient_id = :patientId) av ON sv.id = av.scheduled_vaccine_id
-        WHERE sv.category = :category
-        GROUP BY sv.id
-        ORDER BY sv.index, max(sv.label), max(sv.schedule);
-      `,
-      {
-        replacements: {
-          patientId: req.params.id,
-          category: VACCINE_CATEGORIES.ROUTINE,
-        },
-        model: req.models.ScheduledVaccine,
-        mapToModel: true,
-        type: QueryTypes.SELECT,
-      },
-    );
-
-    const vaccines = results
-      .map(s => s.get({ plain: true }))
-      .reduce((allVaccines, vaccineSchedule) => {
-        const administered = asRealNumber(vaccineSchedule.administered) > 0;
-        if (!allVaccines[vaccineSchedule.label]) {
-          delete vaccineSchedule.administered;
-          vaccineSchedule.schedules = [];
-          allVaccines[vaccineSchedule.label] = vaccineSchedule;
-        }
-        // Exclude historical schedules unless administered
-        if (vaccineSchedule.visibilityStatus !== VISIBILITY_STATUSES.HISTORICAL || administered) {
-          allVaccines[vaccineSchedule.label].schedules.push({
-            schedule: vaccineSchedule.schedule,
-            scheduledVaccineId: vaccineSchedule.id,
-            administered,
-          });
-        }
-        return allVaccines;
-      }, {});
-
-    // Exclude vaccines that already have all the schedules administered for the patient
-    const availableVaccines = Object.values(vaccines).filter(v =>
-      v.schedules.some(s => !s.administered),
-    );
-
-    // Todo: remove this mock logic after NASS-1146 is done
-    const mockedResults = availableVaccines.map((record, index) => {
-      // Get start of week
-      const dueDate = mockDueDate(index);
-      return { ...record, dueDate };
-    });
-    res.send({ count: availableVaccines.length, data: mockedResults });
-  }),
-);
 
 patientVaccineRoutes.get(
   '/:id/scheduledVaccines',
